refactor: 임시 ReservationType을 실제 ReservationType으로 적용
